Split step attribute parsing
Move shared attribute parsing into its own module to preserve
headroom in the step-macro entry module.
